requirements.txt
alembic==1.12.1
fastapi==0.109.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic>=2.5.2
pydantic-settings==2.2.1
python-dotenv>=1.0.0
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.2.1
requests==2.32.3
typer==0.9.0
uvicorn==0.24.0
stripe==11.4.1
celery==5.3.4
redis==5.0.1
pymilvus>=2.6.9
openai>=1.6.0
email-validator>=2.0.0
google-genai>=1.62.0